requirements.txt
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0
pytest>=7.3.0
neuraloperator>=0.3.0

# Optional for Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.23.0
